pyproject.toml
[tool.poetry]
name = "demoshop"
version = "0.1.0"
description = ""
authors = ["Kornilin5 <sasha_kornilin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selene = "2.0.0rc7"
pytest = "^8.1.1"
allure-pytest = "^2.13.4"
requests = "^2.31.0"
curlify = "^2.2.1"
pytest-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
